refactor: externalizes stats
